Whee, 64bit cpus and sizeof.

+++ libs/util/zone.c
@@ -263,8 +263,8 @@ Z_Print (memzone_t *zone)
 
 	for (block = zone->blocklist.next ; ; block = block->next) {
 		Sys_Printf ("block:%p    size:%7i    tag:%3i ofs:%d\n",
-					block, block->size - sizeof (memblock_t) - 4, block->tag,
-					(int) ((byte *) block - (byte *) zone));
+					block, block->size - (int) sizeof (memblock_t) - 4,
+					block->tag, (int) ((byte *) block - (byte *) zone));
 
 		if (block->next == &zone->blocklist)
 			break;			// all blocks have been hit
